logrus: exit on fatal error
The proxy logger now supports an option to exit the process on fatal
errors and panic on panic calls. This is useful for applications that
need to ensure immediate termination on critical errors.

Previously it was enabled by default, this now changes and it must be
explicitly enabled via configuration.