sinit: add test that ensures init twice errors
This commit adds a test that ensures that when `InitializeLogging`
is called twice there is an error. The current implementation
is using a `sync.Once()` so init will only have an effect.

It is document that subsequent calls have no effect but still
I think we need to error, I don't see a use-case where init
with a changed configuration again would not warant an error.