splunk: add missing `defer srv.Close()`
This commit adds missing `defer srv.Close()` in the client
tests for splunk. Without that we leave go-routines for
the test servers behind. This also closes the splunkLogger
when its done (instead of defering it), without that we
would close the http server while there is still a connection
active. As a side-effect the test runs in 4s for me now instead
of 8s.