splunk: avoid race condition between close and flush
When the client is being closed and flush attempt is made, it can lead
to writing to a closed channel, which causes a panic. This change simply
makes the race condition detector happy about it.

It does not change the behavior of the code, when the client is closed,
flush will eventually lead to panic due to closed channel. This behavior
is aligned with the behavior of logging as well. No data must be sent
after the client is closed.