sinit: avoid race conditions during init/flush/close
There is a race condition between InitializeLogger and Flush functions.
This patch fixes this and also introduces function Close that is
blocking which is much more useful than Flush which does not block for
all providers.

The newly introduced Close() function also requires a timeout parameter
for full control of the shutdown process and returns an error when
timeout is reached.

The Close() function also closes the standard logger and sets logrus
proxy and strc to noop outputs.