echo: support context passing
The echo logger proxy did not support contexts, therefore it was not
possible to create per-request loggers that carry context into the slog
handler. With this option, it becomes possible to create loggers that
include request-specific information, such as user IDs or request IDs,
making it easier to trace logs back to specific requests.