strc: native echo middleware
The tracing and logging middleware was only implemented for the standard
Go HTTP library and it was supposed to be used with
`echo.WrapMiddleware` to wrap the Echo handlers. This commit adds a
native Echo middleware implementation that can be used directly with
Echo's `Use` method.

The reason for this change is that echo handler/middleware functions may
return errors and these are processed after all middleware was called.
This caused issues with error handling and logging, as the original
middleware implementation was not able to properly capture and log
errors that occurred in the handler functions.