strc: add ability to override time in Span.Event and Span.End
Span start, event and end time is automatically taken via `time.Now()`
call but there are some use cases when this needs to be overridden to a
specific time. Use special attributes to do that:

    span := strc.Start(ctx, "span name", "started", time.Now())
    span.Event("an event", "at", time.Now())
    span.End("finished", time.Now())

Also adds TraceID() method to Span to return the trace ID. This will be
usefil in images/ library to collect logs from JSONSeq osbuild monitor and
keep them in a LRU cache.